AuthorDTO 수정. Author Entity와 DB간 차이(author_name) 으로 인한 500에러 수정

+++ src/main/java/com/example/project/emotionCore/dto/AuthorDTO.java
@@ -2,12 +2,13 @@ package com.example.project.emotionCore.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
+import lombok.ToString;
 
 @Schema(description = "작가들의 정보")
 @Data
 public class AuthorDTO {
     @Schema(description = "index", example = "1")
-    private int id;
+    private Long id;
 
     @Schema(description = "작가 닉네임", example = "닉네임이다...!")
     private String name;
